requirements.txt
python-dotenv==0.7.1
discord.py==0.16.12
